+++ external/java/src/main/java/ch/epfl/dedis/lib/omniledger/darc/Identity.java
@@ -1,4 +1,4 @@
-package ch.epfl.dedis.lib.darc;
+package ch.epfl.dedis.lib.omniledger.darc;
 
 import ch.epfl.dedis.proto.DarcProto;
