+++ external/java/src/main/java/ch/epfl/dedis/lib/omniledger/darc/Signature.java
@@ -1,4 +1,4 @@
-package ch.epfl.dedis.lib.darc;
+package ch.epfl.dedis.lib.omniledger.darc;
 
 import ch.epfl.dedis.proto.DarcProto;
 import com.google.protobuf.ByteString;
